+++ NEWS.md
@@ -1,6 +1,8 @@
 Version 0.6.7
- - fix: debugger won't exit when the program has exited.
+ - fix: messages send to the gdb process's standard error are not received
  - adjust: the max value of the debug console's vertical scrollbar.
+ - fix: shfit+click not correctly set selection's end
+ - fix: ctrl+home/end not correctly set cursor to start/end of the editor
 
 Version 0.6.6
  - fix: crash when create new file